- 코드 정리
git-svn-id: svn://www.hshgames.com/project_s@140 e3c12c2a-3f9b-4cee-bc1f-3a9786f174a7

+++ proj.win32/Classes/MyUI/ScoreUI.cpp
@@ -53,7 +53,7 @@ bool CScoreUI::initVariable()
 void CScoreUI::insertComma(const char* valueStr, char* resultStr)
 {
 	int len;
-	len = strlen(valueStr);
+	len = static_cast<int>(strlen(valueStr));
 
 	while ((*resultStr++ = *valueStr++)) {
 		if (--len && (len % 3) == 0)
